Refactor types: remove types.hpp, split into separate files; change hitbox offset to 16-bit; delete copy constructors; add initial test code; add CMakeLists.txt for mgc_c and mgc_cpp

+++ src/mgc/components/hitbox.c
@@ -6,7 +6,7 @@
  */
 #include "hitbox.h"
 
-void hitbox_init(mgc_hitbox_t *hitbox, mgc_id_t id, int8_t x0_ofs, int8_t y0_ofs, uint16_t width, uint16_t height) {
+void hitbox_init(mgc_hitbox_t *hitbox, mgc_id_t id, int16_t x0_ofs, int16_t y0_ofs, uint16_t width, uint16_t height) {
     if ( hitbox == NULL ) {
         MGC_WARN("Invalid handler");
         return;
